Node Builder V0.0.3
Switched over to Django as I want this to run online.
Part of me wishes I'd done it sooner but hey now I know how Psycopg works.

Still needs a form and a function call to 'update', but other than that, general idea is that:
there are objects, types, and subtypes.  these inherit one another in postgresql.

So far I've coded the basics of one, and I still need to include /post/ methods